+++ alias/src/intrinsic_sampling_method.py
@@ -21,9 +21,10 @@ import scipy as sp
 import scipy.constants as con
 
 from .utilities import (
-	numpy_remove, bubble_sort, load_npy, make_hdf5, load_hdf5,
-	shape_check_hdf5, save_hdf5
+    numpy_remove, bubble_sort
 )
+from alias.io.hdf5_io import make_hdf5, load_hdf5, save_hdf5, shape_check_hdf5
+from alias.io.numpy_io import load_npy
 
 
 def check_uv(u, v):
